Move scripts in this repo from 'tmux'
The 'tmux' repo is ment to hold dotfiles only. Move scripts here and
keep everything clean and tidy.

Note: The 'app_env' is just a sketch for 'tmux' panes and windows from a Rails
project that should launch multiple engines in different panes and in
the end target the git directory '/application'.